requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
idna==3.6
pipenv==2023.11.15
platformdirs==4.1.0
psycopg2==2.9.9
pycoingecko==3.1.0
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.2.1
virtualenv==20.25.0
